Add files via upload
this class can extract data from a txt file. I've also provided a txt file to test with. it is very particular/specific with how the txt file must be set up. I haven't added a way to add to the txt file yet, I'll try to get that added as soon as I can. If you need to store data so that it is not lost when the program is shut down, this class will allow that to happen when it is finished. the top of the class has a few instructions on how to use the various public methods.